+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added `ttempleton\categorygroupsfield\web\assets\sortable\SortableAsset` (includes the jQuery UI Sortable widget)
+
+### Changed
+- Multi-select category groups fields now use the Selectize library with the Drag & Drop plugin
+
 ## 2.0.1 - 2022-11-05
 
 ### Fixed